db changes, set variables in extras controller

+++ db/migrate/20170418151542_create_products.rb
@@ -3,7 +3,7 @@ class CreateProducts < ActiveRecord::Migration[5.0]
     create_table :products do |t|
       t.string :name
       t.string :sku
-      t.integer :daily_price
+      t.float :daily_price
       t.references :category, index: true, foreign_key: true
       t.text :description
       t.text :detailed_description
